refactor: unit tests
