+++ packages/opentelemetry-core/test/context/HttpTraceContext.test.ts
@@ -31,7 +31,7 @@ describe('HttpTraceContext', () => {
     carrier = {};
   });
 
-  describe('inject', () => {
+  describe('.inject()', () => {
     it('should set traceparent header', () => {
       const spanContext: SpanContext = {
         traceId: 'd4cda95b652f4a1592b449d5929fda1b',
@@ -64,7 +64,7 @@ describe('HttpTraceContext', () => {
     });
   });
 
-  describe('extract', () => {
+  describe('.extract()', () => {
     it('should extract context of a sampled span from carrier', () => {
       carrier[TRACE_PARENT_HEADER] =
         '00-0af7651916cd43dd8448eb211c80319c-b7ad6b7169203331-01';
